浮点数

+++ mycaculator/src/calulatorprogram/Get.java
@@ -1,10 +1,11 @@
 package calulatorprogram;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Stack;
 		//»ùÀà£»
 public class Get {
-	public void getresult(ArrayList<Character> list,Indexof index, StringBuilder builder , Stack<Double> stack , Double allresult ,Character ch) {
+	public void getresult(ArrayList<Character> list,Indexof index, StringBuilder builder , Stack<BigDecimal> stack , BigDecimal allresult ,Character ch) {
 		
 		
 	}
