+++ mycaculator/src/calulatorprogram/chiose.java
@@ -1,5 +1,6 @@
 package calulatorprogram;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Stack;
 
@@ -18,7 +19,7 @@ public class chiose {
 		}
 		return chose;
 	}
- public static void choseGetresult(Get get,ArrayList<Character> list,Indexof index, StringBuilder builder , Stack<Double> stack , Double allresult ,Character ch) {
+ public static void choseGetresult(Get get,ArrayList<Character> list,Indexof index, StringBuilder builder , Stack<BigDecimal> stack , BigDecimal  allresult ,Character ch) {
 	 get.getresult(list, index, builder, stack, allresult, ch);
  }
